% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/merge.flo.r
\name{merge.flo}
\alias{merge.flo}
\title{merge.flo}
\usage{
\method{merge}{flo}(
  stn,
  hy,
  all.lab = F,
  all.flo = F,
  na.rm = T,
  flowData = NULL,
  labData = NULL
)
}
\arguments{
\item{stn}{}

\item{hy}{}

\item{all.lab}{}

\item{all.flo}{}

\item{na.rm}{}

\item{flowData}{}

\item{labData}{}
}
\value{

}
\description{
merge.flo
}
