% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/read.flo.r
\name{read.flo}
\alias{read.flo}
\title{read.flo}
\usage{
read.flo(stn, hy, ttshome = getTTSenv("TTSHOME"), fileLoc = NULL)
}
\arguments{
\item{stn}{}

\item{hy}{}

\item{ttshome}{}

\item{fileLoc}{}
}
\value{

}
\description{
read.flo
}
