% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.lab.r
\name{read.lab}
\alias{read.lab}
\title{read.lab}
\usage{
read.lab(stn, hy, ttshome = getTTSenv("TTSHOME"), fileLoc = NULL)
}
\arguments{
\item{stn}{}

\item{hy}{}

\item{ttshome}{}

\item{fileLoc}{}
}
\value{

}
\description{
read.lab
}
\examples{

}
