% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flo.r
\name{read_flo}
\alias{read_flo}
\title{read_flo}
\usage{
read_flo(stn, hy, ttshome = getTTSenv("TTSHOME"), fileLoc = NULL)
}
\arguments{
\item{stn}{}

\item{hy}{}

\item{ttshome}{}

\item{fileLoc}{}
}
\value{

}
\description{
read_flo
}
