% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lab.r
\name{read_lab}
\alias{read_lab}
\title{read_lab}
\usage{
read_lab(stn, hy, ttshome = getTTSenv("TTSHOME"), fileLoc = NULL)
}
\arguments{
\item{stn}{}

\item{hy}{}

\item{ttshome}{}

\item{fileLoc}{}
}
\value{

}
\description{
read_lab
}
\examples{

}
